revert encoding to T1
I've tried to use texstudio ubuntu 20.4 which is a fairly common
setup and it didn't work with luatex or pdflatex. This might be
related to https://github.com/joaomlourenco/novathesis/issues/133
but I didn't have time to delve deeper into the subject due to lack
of time. The PDF generation should work on common systems without a
lot of effort so I've reverted to T1 and lmodern which also works
with sz on my system.